Move PPO in the example to a separate file
